Connect ESP32 to Adafruit IO
This is a .ino file, to be run on an ESP32. It first connects to WiFi and then to your Adafruit account. It receives data from there and can easily be modified according to your need.  One may also use different display methods on the Adafruit page (https://io.adafruit.com/) suited to the kind of data dealt with.